fix bug where whitespace around zone setting entries prevents save to file and some javascript cleanup
