+++ src/main/resources/templates/zone2.html
@@ -6,7 +6,7 @@
 </head>
 <body>
 <script src="http://ajax.googleapis.com/ajax/libs/jquery/1.9.1/jquery.min.js"></script>
-<script src="commonZoneFun.js"></script>
+<script src="commonZoneDataFun.js"></script>
 
 <h2>
     <pre>      OpenHVACControl - Zone 2</pre>
@@ -15,24 +15,24 @@
 
 <div>
     <h3>
-        Zone 2 &nbsp&nbsp: <textarea id="#ZONEALIAS" style="font-size:12pt; height:25px; width:150px;"></textarea>
+        Zone 2 &nbsp&nbsp: <textarea id="#ZONE2ALIAS" style="font-size:12pt; height:25px; width:150px;"></textarea>
         &nbspSelect:
         <br>
-        Mode &nbsp&nbsp&nbsp&nbsp: <textarea id="#ZONEMODE" style="font-size:12pt; height:25px; width:150px;"></textarea>&nbsp
+        Mode &nbsp&nbsp&nbsp&nbsp: <textarea id="#ZONE2MODE" style="font-size:12pt; height:25px; width:150px;"></textarea>&nbsp
         <button onClick="setMode(2, 'HEAT');" style="font-size:12pt; height:25px; width:100px;">HEAT</button>
         <button onClick="setMode(2, 'COOL');" style="font-size:12pt; height:25px; width:100px;">COOL</button>
         <button onClick="setMode(2, 'OFF');" style="font-size:12pt; height:25px; width:100px;">OFF</button>
         <br>
-        Temp &nbsp&nbsp&nbsp&nbsp&nbsp: <textarea id="#ZONEUSINGTEMP" style="font-size:12pt; height:25px; width:150px;"></textarea>
+        Temp &nbsp&nbsp&nbsp&nbsp&nbsp: <textarea id="#ZONE2USINGTEMP" style="font-size:12pt; height:25px; width:150px;"></textarea>
         <br>
-        Setpoint : <textarea id="#ZONESETPOINT" style="font-size:12pt; height:25px; width:150px;"></textarea>
+        Setpoint : <textarea id="#ZONE2SETPOINT" style="font-size:12pt; height:25px; width:150px;"></textarea>
         <button onClick="increaseSP(2);" style="font-size:12pt; height:25px; width:100px;">&nbsp;Up&nbsp;</button>
         <button onClick="decreaseSP(2);" style="font-size:12pt; height:25px; width:100px;">Down</button>
         <br>
-        Fan &nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp: <textarea id="#ZONEFAN" style="font-size:12pt; height:25px; width:150px;"></textarea>
+        Fan &nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp: <textarea id="#ZONE2FAN" style="font-size:12pt; height:25px; width:150px;"></textarea>
         <button onClick="toggleFanOnAuto(2);" style="font-size:12pt; height:25px; width:100px;">&nbsp;Auto/On</button>
         <br>
-        &nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp <textarea id="#OCCUPANCY" style="font-size:12pt; height:25px; width:150px;"></textarea>
+        &nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp&nbsp <textarea id="#ZONE2OCCUPANCY" style="font-size:12pt; height:25px; width:150px;"></textarea>
         <button onClick="setOccupancy(2, true);" style="font-size:12pt; height:25px; width:150px;">&nbsp;OCCUPIED</button>
         <button onClick="setOccupancy(2, false);" style="font-size:12pt; height:25px; width:150px;">&nbsp;UNOCCUPIED</button>
     </h3>
@@ -47,10 +47,10 @@
 
 <script type="text/javascript">
     $( document ).ready(function() {
-        getSettings(2);
+        getZoneData(2);
     });
 
-    var update_loop = setInterval(getSettings(2), 2000);
+    var update_loop = setInterval(getZoneData(2), 2000);
 
 </script>
